refactor(MandateLanguage): Change format of MandateLanguage Enum
