+++ src/main/scala/com/particeep/api/models/wallet/sepa/SddMandateCreation.scala
@@ -2,7 +2,7 @@ package com.particeep.api.models.wallet.sepa
 
 import play.api.libs.json.{ Json, OFormat }
 
-import com.particeep.api.models.enums.MandateLanguage.MandateLanguage
+import com.particeep.api.models.enums.MandateLanguage
 
 case class SddMandateCreation(
   owner_ip:         String                  = "",
